Improvements to transactions package
