+++ bank/transactions_show/transactions_show.go
@@ -2,12 +2,12 @@ package transactions_show
 
 import "github.com/luistm/banksaurus/bank"
 
-// New creates a transactions show interactor
+// New creates a transaction show interactor
 func New() (bank.Interactor, error) {
 	return &TransactionsShow{}, nil
 }
 
-// TransactionsShow shows transactions
+// TransactionsShow shows transaction
 type TransactionsShow struct{}
 
 // Execute ...
