+++ cmd/banksaurus/banksaurusCLIacceptance_test.go
@@ -106,7 +106,7 @@ func TestSystem(t *testing.T) {
 			expected: "COMPRA CONTINENTE MAI\nCOMPRA FARMACIA SAO J\n",
 		},
 		{
-			name:     "Show transactions, from the records file just loaded",
+			name:     "Show transaction, from the records file just loaded",
 			command:  []string{"transaction", "show"},
 			expected: "COMPRA CONTINENTE MAI 77.52\nCOMPRA FARMACIA SAO J 95.09",
 		},
