+++ lib/transaction/entities_test.go
@@ -1,11 +1,11 @@
-package transactions_test
+package transaction_test
 
 import (
 	"testing"
 
 	"github.com/luistm/banksaurus/elib/testkit"
 	"github.com/luistm/banksaurus/lib/sellers"
-	"github.com/luistm/banksaurus/lib/transactions"
+	"github.com/luistm/banksaurus/lib/transaction"
 	"github.com/shopspring/decimal"
 )
 
@@ -14,7 +14,7 @@ func TestUnitTransactionNew(t *testing.T) {
 	s := sellers.New("TheSellerSlug", "TheSellerName")
 	value := "1.1"
 
-	tr, err := transactions.New(s, value)
+	tr, err := transaction.New(s, value)
 
 	testkit.AssertIsNil(t, err)
 	testkit.AssertEqual(t, tr.Seller, s)
@@ -22,7 +22,7 @@ func TestUnitTransactionNew(t *testing.T) {
 	testkit.AssertIsNil(t, err)
 	testkit.AssertEqual(t, tr.Value().String(), v.String())
 
-	tr, err = transactions.New(s, "")
+	tr, err = transaction.New(s, "")
 
 	testkit.AssertIsNil(t, err)
 	testkit.AssertEqual(t, tr.Seller, s)
